misc/HtmlTestSuite/results/test-0962.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
<head>
<title>Inline classic scripts with nomodule content attribute must not run</title>
<link href="mailto:utatane.tea@gmail.com" rel="author" title="Yusuke Suzuki"/>
<link href="mailto:rniwa@webkit.org" rel="author" title="Ryosuke Niwa"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
</head>
<body>
<script>
window.executed = true;
</script>
<script>

test(() =&gt; {
    assert_true(executed);
}, 'An inline classic script without nomodule content attribute must run');


window.executed = false;
</script>
<script nomodule="">
window.executed = true;
</script>
<script>

test(() =&gt; {
    assert_false(executed);
}, 'An inline classic script with nomodule content attribute must not run');

</script>
<script>

test(() =&gt; {
    window.executed = false;
    const element = document.createElement("script");
    element.noModule = false;
    element.textContent = `window.executed = true`;
    document.body.appendChild(element);
    assert_true(window.executed);
}, 'An inline classic script element dynamically inserted after noModule was set to false must run.');

test(() =&gt; {
    window.executed = false;
    const element = document.createElement("script");
    element.noModule = true;
    element.textContent = `window.executed = true`;
    document.body.appendChild(element);
    assert_false(window.executed);
}, 'An inline classic script element dynamically inserted after noModule was set to true must not run.');

window.executed = false;
</script>
</body>
</html>
